refactor(PropertyTypeIterable): remove deprecated PropertyTypeArray class
