+++ src/Metadata/PropertyTypeIterable.php
@@ -12,8 +12,18 @@ use Doctrine\Common\Collections\Collection;
  *  - we're not merging a plain array PropertyTypeIterable into a hashmap one,
  *  - and the collection classes of each are either not present on both sides, or are the same, or parent-child of one another
  */
-final class PropertyTypeIterable extends PropertyTypeArray
+final class PropertyTypeIterable extends AbstractPropertyType
 {
+    /**
+     * @var PropertyType
+     */
+    protected $subType;
+
+    /**
+     * @var bool
+     */
+    protected $hashmap;
+
     /**
      * @var string
      */
@@ -24,8 +34,10 @@ final class PropertyTypeIterable extends PropertyTypeArray
      */
     public function __construct(PropertyType $subType, bool $hashmap, bool $nullable, string $traversableClass = null)
     {
-        parent::__construct($subType, $hashmap, $nullable, null != $traversableClass);
+        parent::__construct($nullable);
 
+        $this->subType = $subType;
+        $this->hashmap = $hashmap;
         $this->traversableClass = $traversableClass;
     }
 
@@ -41,25 +53,20 @@ final class PropertyTypeIterable extends PropertyTypeArray
             $array .= sprintf('|\\%s<%s%s>', $this->traversableClass, $this->subType, $collectionType);
         }
 
-        return ((string) $this->subType).$array.AbstractPropertyType::__toString();
+        return ((string) $this->subType).$array.parent::__toString();
+    }
+
+    public function isHashmap(): bool
+    {
+        return $this->hashmap;
     }
 
     /**
-     * @deprecated Please prefer using {@link getTraversableClass}
-     *
-     * @return class-string<Collection>|null
+     * Returns the type of the next level, which could be an array or hashmap or another type.
      */
-    public function getCollectionClass(): ?string
+    public function getSubType(): PropertyType
     {
-        return $this->isCollection() ? null : $this->traversableClass;
-    }
-
-    /**
-     * @deprecated Please prefer using {@link isTraversable}
-     */
-    public function isCollection(): bool
-    {
-        return (null != $this->traversableClass) && is_a($this->traversableClass, Collection::class, true);
+        return $this->subType;
     }
 
     /**
@@ -79,6 +86,19 @@ final class PropertyTypeIterable extends PropertyTypeArray
         return null != $this->traversableClass;
     }
 
+    /**
+     * Goes down the type until it is not an array or hashmap anymore.
+     */
+    public function getLeafType(): PropertyType
+    {
+        $type = $this->getSubType();
+        while ($type instanceof self) {
+            $type = $type->getSubType();
+        }
+
+        return $type;
+    }
+
     public function merge(PropertyType $other): PropertyType
     {
         $nullable = $this->isNullable() && $other->isNullable();
@@ -88,9 +108,9 @@ final class PropertyTypeIterable extends PropertyTypeArray
             return new self($this->subType, $this->isHashmap(), $nullable, $thisTraversableClass);
         }
         if ($this->isTraversable() && (($other instanceof PropertyTypeClass) && is_a($other->getClassName(), \Traversable::class, true))) {
-            return new self($this->getSubType(), $this->isHashmap(), $nullable, $this->findCommonCollectionClass($thisTraversableClass, $other->getClassName()));
+            return new self($this->getSubType(), $this->isHashmap(), $nullable, $this->findCommonTraversableClass($thisTraversableClass, $other->getClassName()));
         }
-        if (!$other instanceof parent) {
+        if (!$other instanceof self) {
             throw new \UnexpectedValueException(sprintf('Can\'t merge type %s with %s, they must be the same or unknown', self::class, \get_class($other)));
         }
 
@@ -105,7 +125,7 @@ final class PropertyTypeIterable extends PropertyTypeArray
 
         $otherTraversableClass = $other->isTraversable() ? $other->getTraversableClass() : null;
         $hashmap = $this->isHashmap() || $other->isHashmap();
-        $commonClass = $this->findCommonCollectionClass($thisTraversableClass, $otherTraversableClass);
+        $commonClass = $this->findCommonTraversableClass($thisTraversableClass, $otherTraversableClass);
 
         if ($other->getSubType() instanceof PropertyTypeUnknown) {
             return new self($this->getSubType(), $hashmap, $nullable, $commonClass);
@@ -121,7 +141,7 @@ final class PropertyTypeIterable extends PropertyTypeArray
      * Find the most derived class that doesn't deny both class hints, meaning the most derived
      * between left and right if one is a child of the other
      */
-    private function findCommonCollectionClass(?string $left, ?string $right): ?string
+    private function findCommonTraversableClass(?string $left, ?string $right): ?string
     {
         if (null === $right) {
             return $left;
